requirements.txt
# ============================================
# WeBoostX 2.0 - Python Dependencies
# ============================================

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Validation & Settings
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Scheduler
apscheduler==3.10.4

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# LINE Notify
requests==2.31.0

# Development
python-dotenv==1.0.1

